backend/requirements.txt
# BugYou Backend Requirements
# Core web framework
flask==3.0.0
flask-cors==4.0.0

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.25

# Authentication & Security
bcrypt==4.1.2
PyJWT==2.8.0
flask-jwt-extended==4.6.0

# Environment management
python-dotenv==1.0.0

# HTTP requests (for Piston API)
requests==2.31.0

# Data validation
marshmallow==3.20.2

# Development tools
pytest==7.4.4
pytest-flask==1.3.0

# Optional: If using async features
# aiohttp==3.9.1
# asyncpg==0.29.0 